Update README.md
# FileExtensionVirus
A project work made with Java, The application is made to secretly change any file extension to a desire extension. After a success process, a Readme.txt field will be generated and distributed to all directories of your c:\ or root directory. This application works if only the user launch it. It works with no GUI nor cmd interface. After all, the Readme.txt file generated will instruct the user to contact the developer of the system to send the file decryptor or extension remover. Note the User can manually edit the extension and the files will not lose its content or corrupt. But it would be a tedious work doing it manually for the all the files in the entire system.
The extension could be any short letters preferred. (like  . ogudu,  .owogo, .ohgha  .emls).

This is for educational purposes, only use it for a permitted system. Using this software on a system without the person of the system owner is illegal.  You can modify this software to work perfectly for you, redistribute. 

The software is under development. Working on how to render the file corrupted if a user manual change the extension created by the software.